Extended the EditPointForms class from the new Smart class, the common methods from the EditPointForm were moved to the Smart
